+++ pkgs/ffigen/lib/src/code_generator/objc_built_in_functions.dart
@@ -76,6 +76,7 @@ class ObjCBuiltInFunctions {
     'NSNotification',
     'NSNumber',
     'NSObject',
+    'NSOrderedCollectionDifference',
     'NSOrderedSet',
     'NSOutputStream',
     'NSProxy',