+++ pkgs/ffigen/lib/src/header_parser/translation_unit_parser.dart
@@ -7,12 +7,10 @@ import 'package:logging/logging.dart';
 import '../code_generator.dart';
 import 'clang_bindings/clang_bindings.dart' as clang_types;
 import 'data.dart';
-import 'includer.dart';
 import 'sub_parsers/functiondecl_parser.dart';
 import 'sub_parsers/macro_parser.dart';
 import 'sub_parsers/objcinterfacedecl_parser.dart';
 import 'sub_parsers/objcprotocoldecl_parser.dart';
-import 'sub_parsers/typedefdecl_parser.dart';
 import 'sub_parsers/var_parser.dart';
 import 'type_extractor/extractor.dart';
 import 'utils.dart';
@@ -35,6 +33,7 @@ Set<Binding> parseTranslationUnit(clang_types.CXCursor translationUnitCursor) {
           case clang_types.CXCursorKind.CXCursor_UnionDecl:
           case clang_types.CXCursorKind.CXCursor_EnumDecl:
           case clang_types.CXCursorKind.CXCursor_ObjCInterfaceDecl:
+          case clang_types.CXCursorKind.CXCursor_TypedefDecl:
             addToBindings(bindings, _getCodeGenTypeFromCursor(cursor));
             break;
           case clang_types.CXCursorKind.CXCursor_ObjCCategoryDecl:
@@ -48,11 +47,6 @@ Set<Binding> parseTranslationUnit(clang_types.CXCursor translationUnitCursor) {
             break;
           case clang_types.CXCursorKind.CXCursor_VarDecl:
             addToBindings(bindings, parseVarDeclaration(cursor));
-            break;
-          case clang_types.CXCursorKind.CXCursor_TypedefDecl:
-            if (config.includeUnusedTypedefs) {
-              addToBindings(bindings, parseTypedefDeclaration(cursor));
-            }
             break;
           default:
             _logger.finer('rootCursorVisitor: CursorKind not implemented');
@@ -80,7 +74,7 @@ void addToBindings(Set<Binding> bindings, Binding? b) {
 }
 
 BindingType? _getCodeGenTypeFromCursor(clang_types.CXCursor cursor) {
-  final t = getCodeGenType(cursor.type(), ignoreFilter: false);
+  final t = getCodeGenType(cursor.type());
   return t is BindingType ? t : null;
 }
 
@@ -95,4 +89,21 @@ void buildUsrCursorDefinitionMap(clang_types.CXCursor translationUnitCursor) {
       rethrow;
     }
   });
+}
+
+/// True if a cursor should be included based on headers config, used on root
+/// declarations.
+bool shouldIncludeRootCursor(String sourceFile) {
+  // Handle empty string in case of system headers or macros.
+  if (sourceFile.isEmpty) {
+    return false;
+  }
+
+  // Add header to seen if it's not.
+  if (!bindingsIndex.isSeenHeader(sourceFile)) {
+    bindingsIndex.addHeaderToSeen(
+        sourceFile, config.shouldIncludeHeader(Uri.file(sourceFile)));
+  }
+
+  return bindingsIndex.getSeenHeaderStatus(sourceFile)!;
 }